third_party/WebKit/LayoutTests/typedcssom/inlinestyle/properties/border-bottom-width-expected.txt

This is a testharness.js-based test.
PASS Setting border-bottom-width to thin 
PASS Setting border-bottom-width to medium 
PASS Setting border-bottom-width to thick 
PASS Setting border-bottom-width to initial 
PASS Setting border-bottom-width to inherit 
PASS Setting border-bottom-width to unset 
FAIL Setting border-bottom-width to CSSUnitValue with value 1px Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 3em Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 4ex Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 5ch Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 6rem Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 7vw Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 8vh Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 9vmin Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 10vmax Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 11cm Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 12mm Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 13in Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 14pc Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Setting border-bottom-width to CSSUnitValue with value 15pt Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting border-bottom-width to invalid value CSSUnitValue throws 
PASS Setting border-bottom-width to invalid value null throws 
PASS Setting border-bottom-width to invalid value undefined throws 
PASS Setting border-bottom-width to invalid value true throws 
PASS Setting border-bottom-width to invalid value false throws 
PASS Setting border-bottom-width to invalid value 1 throws 
PASS Setting border-bottom-width to invalid value hello throws 
PASS Setting border-bottom-width to invalid value [object Object] throws 
PASS Setting border-bottom-width to invalid value CSSKeywordValue throws 
PASS Getting border-bottom-width when it is set to thin 
PASS Getting border-bottom-width when it is set to medium 
PASS Getting border-bottom-width when it is set to thick 
PASS Getting border-bottom-width when it is set to initial 
PASS Getting border-bottom-width when it is set to inherit 
PASS Getting border-bottom-width when it is set to unset 
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 1px assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 3em assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 4ex assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 5ch assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 6rem assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 7vw assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 8vh assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 9vmin assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 10vmax assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 11cm assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 12mm assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 13in assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 14pc assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting border-bottom-width with a CSSUnitValue whose value is 15pt assert_equals: typeof result expected "CSSUnitValue" but got "CSSStyleValue"
FAIL getAll for single-valued border-bottom-width assert_equals: Returned type is incorrect: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Delete border-bottom-width removes the value from the styleMap 
FAIL border-bottom-width shows up in getProperties Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting border-bottom-width to a sequence throws 
PASS Appending to border-bottom-width throws 
Harness: the test ran to completion.
